+++ src/app/auth-callback/page.tsx
@@ -2,7 +2,7 @@
 
 import { Loader2 } from "lucide-react"
 import { useRouter, useSearchParams } from "next/navigation"
-import { trpc } from "../_trpc/client"
+import { trpc } from "@/app/_trpc/client"
 
 
 function Page() {