+++ src/app/layout.tsx
@@ -4,6 +4,7 @@ import { cn } from '@/lib/utils'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import './globals.css'
+import 'react-loading-skeleton/dist/skeleton.css'
 
 const inter = Inter({ subsets: ['latin'] })