+++ changes.md
@@ -4,7 +4,7 @@ Note: To be able to build this app exactly as intended, you can use Josh's [`pnm
 
 If you want to build using more recent/up-to-date modules, you can follow along here.
 
-## Creating Our Dashboard [1:27:12](https://www.youtube.com/watch?v=ucX2zXAZ1I0&t=5233s)
+## Creating Our Dashboard [[1:27:12](https://www.youtube.com/watch?v=ucX2zXAZ1I0&t=5233s)]
 ### Changes to Kinde's `getUser()`
 
 
@@ -17,5 +17,11 @@ const user = getUser()
 NEW VERSION:
 const user = await getUser()
 ```
+
+By [2:18:29](https://www.youtube.com/watch?v=ucX2zXAZ1I0&t=8309s) I had to add `await getUser()` on:
+- `src/app/dashboard/page.tsx` - come back and link these later
+- `src/trpc/index.ts` - come back and link these later
+- [2:44:58] `src/trpc/trpc.ts` 
+
 
 Note: This also means you need to update certain functions like `src/app/dashboard/page.tsx` to `async` when you call await inside.